requirements.txt
Flask
PyJWT
opencv-python
Pillow
requests